+++ wb/lib/wb/src/wb_merep.h
@@ -35,8 +35,8 @@ public:
   wb_tdrep *tdrep( pwr_tStatus *sts, pwr_tTid tid);
   wb_tdrep *tdrep( pwr_tStatus *sts, wb_name name);
 
-  int getAttrInfoRec( wb_attrname *attr, pwr_eBix bix, pwr_tCid cid, int *size,
-                      int *offset, pwr_tTid *tid, int *elements, 
+  int getAttrInfoRec( wb_attrname *attr, pwr_eBix bix, pwr_tCid cid, size_t *size,
+                      size_t *offset, pwr_tTid *tid, int *elements, 
                       pwr_eType *type, int *flags, int level);
 
 };
